bug in timer code, signed int priority caused overflow + deadlock

+++ src/kernel/include/datastructures.h
@@ -4,7 +4,7 @@
 #include "macros.h"
 
 typedef struct pqnode_s{
-    int priority;
+    uint64_t priority;
     uintptr_t element; // generic element pointer
 } pqnode_t;
